update type and file names
