+++ docs/en/integrations/firestore/tools/firestore-mongodb-execute-mql.md
@@ -1,14 +1,14 @@
 ---
-title: "firestoremongodb-execute-mql"
+title: "firestore-mongodb-execute-mql"
 type: docs
 weight: 1
 description: >
-  A "firestoremongodb-execute-mql" tool executes MongoDB Query Language (MQL) queries and aggregation pipelines against Firestore.
+  A "firestore-mongodb-execute-mql" tool executes MongoDB Query Language (MQL) queries and aggregation pipelines against Firestore.
 ---
 
 ## About
 
-A `firestoremongodb-execute-mql` tool executes MongoDB Query Language (MQL) statements or aggregation pipelines against Firestore.
+A `firestore-mongodb-execute-mql` tool executes MongoDB Query Language (MQL) statements or aggregation pipelines against Firestore.
 
 This tool allows data agents and LLMs to run dynamic MQL queries such as aggregation pipelines or document queries directly against Firestore databases.
 
@@ -27,7 +27,7 @@ This tool allows data agents and LLMs to run dynamic MQL queries such as aggrega
 ```yaml
 kind: tool
 name: execute_mql
-type: firestoremongodb-execute-mql
+type: firestore-mongodb-execute-mql
 source: my-firestore-source
 description: Use this tool to execute MQL queries against Firestore.
 ```
@@ -36,6 +36,6 @@ description: Use this tool to execute MQL queries against Firestore.
 
 | **field**   | **type** | **required** | **description** |
 |-------------|:--------:|:------------:|-----------------|
-| type | string | true | Must be "firestoremongodb-execute-mql". |
+| type | string | true | Must be "firestore-mongodb-execute-mql". |
 | source | string | true | Name of the Firestore source to execute queries against. |
 | description | string | true | Description of the tool that is passed to the LLM. |
